Change config for public devnet
Since we now have gas costs for FHE operations, set block gas limit to
100000000 gas. As a reference, an ERC20 transfer estimates to around
1220732 gas for euint32, meaning a block could have up to 100 ERC20
transfers. These values might change as we test more, but we need a
starting point.

Since we now have compact ciphertexts for input, set max block size to
4194304 (4MB).

Disable the `personal` and `debug` namespaces on the JSON-RPC server.

Reinstate the minimum gas price to 0.0001aevmos (as it was before our
FHE modifications).

Put a gas cap of 500000000 gas for `eth_call` (i.e. view functions) over
the JSON-RPC interface. That limits the amount of computation users can
do "for free" over JSON-RPC.

Remove custom gas prices and adjustments that we added initially when
we worked with bigger ciphertexts and proofs.